move into tls module for readability
